Add files via upload
This script uses two repeats are characteristic of the pilE and pilS loci to allows the identification and the extraction pilS loci. 
It works by identifying  specific genetic motifs (cys2 and  SmaCla) in genomic sequences, and extracts the surrounding sequences fragments pilS and pilE. Then the extracted fragments are stored in a fasta file, and a CSV file report is generated.